refactor: share generateSchemaData between the pg and sqlite builders
`generateSchemaData` was duplicated between `src/util/builders/pg.ts` and
`src/util/builders/sqlite.ts` — ~750 lines each that differed in only four
places. A normalized diff of the two bodies came back identical apart from
comment wording, formatting, and SQLite's two synchronous-driver guards.

The shared body now lives in `src/util/builders/schema-data.ts`.
`createSchemaDataGenerator(adapter)` is called once per dialect module at
import time and returns that dialect's implementation, so the write
resolvers — which close over the dialect's primary-key lookup — are still
built once rather than per `buildSchema` call.

A dialect supplies only what it genuinely owns, via `DialectSchemaAdapter`:

- `tableClass` / `getTableConfig` / `primaryKeyPropNames` — how tables are
  picked out of the schema module and how their keys are read;
- `nullOrdering` — PostgreSQL sorts NULLs largest, SQLite smallest, and
  relation pagination has to match;
- `preflight` — optional validation of the handle against the options.
  SQLite uses it for the two guards that reject `features.nestedWrites` and
  `transactions: 'auto'` on a synchronous driver;
- `generateSelectArray` / `generateSelectSingle` / `generateUpdateMany` —
  the three resolver generators that still differ. The select pair is kept
  dialect-local deliberately: pg carries a plain-select fallback for tables
  without RQB support where sqlite throws, and this change is not the place
  to alter that behavior.

MySQL stays off this path: its writes return no rows, so its mutation
fields, return types and resolvers differ throughout.

Also aligns pg's public `generateSchemaData` signature with sqlite's — it
now takes the same `<TDrizzleInstance, TSchema>` pair and returns
`GeneratedEntities<TDrizzleInstance, TSchema>` instead of its own one-generic
form.

Net -670 lines, no behavior change. Verified with `npm run typecheck` (both
projects), biome, the 70-file pglite/sqlite suite (1044 tests) and the
Docker pg/mysql suite (180 tests).
